+++ CHANGELOG.md
@@ -7,10 +7,14 @@ See DataLab [roadmap page](https://datalab-platform.com/en/contributing/roadmap.
 💥 New features and enhancements:
 
 * New image processing features:
-  * Added new "Gaussian frequency filter" feature:
+  * New "Gaussian frequency filter" feature:
     * This feature allows to filter an image in the frequency domain using a Gaussian filter
     * It is implemented in the `sigima.proc.image.freq_fft` function (requires Sigima v0.3)
     * This closes [Issue #205](https://github.com/DataLab-Platform/DataLab/issues/205) - Add Gaussian frequency filter for images
+  * New "Erase area" feature:
+    * This feature allows to erase an area in an image as defined by a region of interest
+    * It is implemented in the `sigima.proc.image.erase` function (requires Sigima v0.3)
+    * This closes [Issue #204](https://github.com/DataLab-Platform/DataLab/issues/204) - Add support for background subtraction and inpainting on ROI
 
 * New signal processing features:
   * Zero padding: